vc-sm: Move headers out of arch directory
